weather-python/requirements.txt
mcp[cli]
httpx
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0